requirements.txt
arabic-reshaper==3.0.0
numpy==1.24.3
pandas==2.0.2
pygame==2.4.0
pynput==1.7.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
screeninfo==0.8.1
six==1.16.0
tzdata==2023.3
